third_party/WebKit/LayoutTests/fast/plugins/plugin-placeholder-not-exposed.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<script>
description('Ensures that the private PluginPlaceholderElement interface is not exposed.');
shouldBeUndefined('window.PluginPlaceholderElement');
shouldThrow('PluginPlaceholderElement');
shouldThrow('PluginPlaceholderElement.prototype');
</script>
